Revert "Clean up flutter_gallery__transition_perf (#63609)"
This reverts commit 360023e2bcb17fa63a563aebe8dc6ef6428c84cd.
