--- dev/integration_tests/flutter_gallery/test_driver/transitions_perf_hybrid_test.dart
@@ -1,9 +1,0 @@
-// Copyright 2014 The Flutter Authors. All rights reserved.
-// Use of this source code is governed by a BSD-style license that can be
-// found in the LICENSE file.
-
-import 'transitions_perf_test.dart' as transitions_perf_test;
-
-void main([List<String> args = const <String>[]]) => transitions_perf_test.main(
-  <String>['--hybrid', ...args],
-);
