+++ dev/integration_tests/flutter_gallery/test_driver/transitions_perf_test.dart
@@ -153,7 +153,6 @@ Future<void> runDemos(List<String> demos, FlutterDriver driver) async {
 
 void main([List<String> args = const <String>[]]) {
   final bool withSemantics = args.contains('--with_semantics');
-  final bool hybrid = args.contains('--hybrid');
   group('flutter gallery transitions', () {
     FlutterDriver driver;
     setUpAll(() async {
@@ -187,11 +186,7 @@ void main([List<String> args = const <String>[]]) {
       // Collect timeline data for just a limited set of demos to avoid OOMs.
       final Timeline timeline = await driver.traceAction(
         () async {
-          if (hybrid) {
-            await driver.requestData('profileDemos');
-          } else {
-            await runDemos(kProfiledDemos, driver);
-          }
+          await runDemos(kProfiledDemos, driver);
         },
         streams: const <TimelineStream>[
           TimelineStream.dart,
@@ -210,12 +205,8 @@ void main([List<String> args = const <String>[]]) {
           histogramPath);
 
       // Execute the remaining tests.
-      if (hybrid) {
-        await driver.requestData('profileDemos');
-      } else {
-        final Set<String> unprofiledDemos = Set<String>.from(_allDemos)..removeAll(kProfiledDemos);
-        await runDemos(unprofiledDemos.toList(), driver);
-      }
+      final Set<String> unprofiledDemos = Set<String>.from(_allDemos)..removeAll(kProfiledDemos);
+      await runDemos(unprofiledDemos.toList(), driver);
 
     }, timeout: const Timeout(Duration(minutes: 5)));
   });
