+++ dev/integration_tests/flutter_gallery/test_driver/transitions_perf_with_semantics.dart
@@ -1,0 +1,9 @@
+// Copyright 2014 The Flutter Authors. All rights reserved.
+// Use of this source code is governed by a BSD-style license that can be
+// found in the LICENSE file.
+
+import 'transitions_perf.dart' as transitions_perf;
+
+void main() {
+  transitions_perf.main();
+}
